requirements.txt
pandas
pandas_ta
python-binance
requests
